refactor: relocate ProjectsMap component
