+++ src/features/projectsV2/ProjectsMap/index.tsx
@@ -2,7 +2,7 @@ import Map from 'react-map-gl-v7/maplibre';
 import 'maplibre-gl/dist/maplibre-gl.css';
 import { NavigationControl } from 'react-map-gl-v7/maplibre';
 import { useRef, MutableRefObject } from 'react';
-import { useProjectsMap } from '../../../projectsV2/ProjectsMapContext';
+import { useProjectsMap } from '../ProjectsMapContext';
 
 function ProjectsMap() {
   const mapRef: MutableRefObject<null> = useRef(null);
